# Conflicts:
#	configIni.php

+++ hserver/dbaccess/db_records.php
@@ -94,13 +94,13 @@
 
         // OWNER -----------
         $ownerid = intval($ownerid);
-        if(!$ownerid && isset($userDefaultOwnerGroupID)){
+        if(!($ownerid>=0) && isset($userDefaultOwnerGroupID)){
             $ownerid = $userDefaultOwnerGroupID;
         }
-        if(!$ownerid){
+        if(!($ownerid>=0)){
             $ownerid = @$sysvals['sys_NewRecOwnerGrpID'];
         }
-        if(!$ownerid){
+        if(!($ownerid>=0)){
             $ownerid = $system->get_user_id();
         }
 
@@ -331,7 +331,7 @@
                 }*/
                 
                 //add reverce field "Parent Entity" (#247) in child resource record
-                if(@$values['dtl_ParentChild']==true){
+                if(@$values['dtl_ParentChild']==true && defined('DT_PARENT_ENTITY')){
                     
                         //find child record
                         list($dtl_ID, $parentID) = mysql__select_array($mysqli,
@@ -586,19 +586,21 @@
     //
     function removeReverseChildToParentPointer($mysqli, $parent_id, $rectype){
 
-       //get list of valid record 
-       $query = 'SELECT dtl_Value FROM recDetails, defRecStructure WHERE dtl_RecID='
-        .$parent_id.' AND dtl_DetailTypeID=rst_DetailTypeID AND rst_CreateChildIfRecPtr=1 AND rst_RecTypeID='.$rectype;
-       
-       $recids = mysql__select_list2($mysqli, $query);
-               
-       $query = 'DELETE FROM recDetails WHERE dtl_Value='.$parent_id.' AND dtl_DetailTypeID='.DT_PARENT_ENTITY;
-       
-       if(is_array($recids) && count($recids)>0){
-            $query = $query.' AND dtl_RecID NOT IN ('.implode(',',$recids).')';
+       if(defined('DT_PARENT_ENTITY')){
+           //get list of valid record 
+           $query = 'SELECT dtl_Value FROM recDetails, defRecStructure WHERE dtl_RecID='
+            .$parent_id.' AND dtl_DetailTypeID=rst_DetailTypeID AND rst_CreateChildIfRecPtr=1 AND rst_RecTypeID='.$rectype;
+           
+           $recids = mysql__select_list2($mysqli, $query);
+                   
+           $query = 'DELETE FROM recDetails WHERE dtl_Value='.$parent_id.' AND dtl_DetailTypeID='.DT_PARENT_ENTITY;
+           
+           if(is_array($recids) && count($recids)>0){
+                $query = $query.' AND dtl_RecID NOT IN ('.implode(',',$recids).')';
+           }
+            
+           $mysqli->query($query);
        }
-        
-       $mysqli->query($query);
     }
 
     // verify ACCESS RIGHTS -------------
@@ -692,7 +694,7 @@
         }
         
         //if not defined set current values
-        if(!$ownerid){
+        if(!($ownerid>=0)){
             $ownerid = $record["rec_OwnerUGrpID"];
         }
         if(!$access){
@@ -794,6 +796,14 @@
         
         $mysqli = $system->get_mysqli();
 
+        if(!(isset($rectype) && $rectype>0)){
+            $rectype = mysql__select_value($mysqli, "select rec_RecTypeID from Records where rec_ID=".$recID);
+            if(!($rectype>0)){
+                $system->addError(HEURIST_DB_ERROR, 'Cannot get record for title mask update. Rec#'.$recID);
+                return false;
+            }
+        }
+        
         $mask = mysql__select_value($mysqli,"select rty_TitleMask from defRecTypes where rty_ID=".$rectype);
         if(!$mask){
             $system->addError(HEURIST_DB_ERROR, 'Cannot get title mask for record type', $mysqli->error);
